add reservation route

+++ cruise-api/routes/api.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\CruiseController;
 use App\Http\Controllers\ParkingController;
 use App\Http\Controllers\ReservationController;
 
+
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -28,41 +29,44 @@ Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
 
 
 //cruise//
-Route::get('/cruise', [CruiseController::class,'get_cruise'] );
+Route::get('/cruise', [CruiseController::class,'get_cruise'] ); 
 Route::get('/Admin', [CruiseController::class,'cruiseAdmin'] )->middleware('auth');
-Route::post('/addCruise', [CruiseController::class,'addCruise'] );
-Route::delete('/destroyCruise/{id}', [CruiseController::class,'destroy'] );
+Route::post('/addCruise', [CruiseController::class,'addCruise'] )->middleware('auth');
+Route::delete('/destroyCruise/{id}', [CruiseController::class,'destroy'] )->middleware('auth');
+Route::get('/show/{id}', [CruiseController::class,'show'] )->middleware('auth');
 
 
 
 
 //Reservation//
-Route::post('/addReservation', [ReservationController::class,'addReservation'] );
-Route::get('/getReservation/{id}', [ReservationController::class,'getReservation'] );
-Route::get('/tickets/{id}', [ReservationController::class,'getReservationsByUserId'] );
-Route::post('/updateReservation/{id}', [ReservationController::class,'updateReservation'] );
+Route::post('/addReservation/{id}', [ReservationController::class,'addReservation'] )->middleware('auth');
+Route::get('/getReservation/{id}', [ReservationController::class,'getReservation'] )->middleware('auth');
+Route::get('/tickets/{id}', [ReservationController::class,'getReservationsByUserId'] )->middleware('auth');
+Route::post('/updateReservation/{id}', [ReservationController::class,'updateReservation'] )->middleware('auth');
 
 
 
 //port//
-Route::get('/AdminP', [PortController::class,'portAdmin'] );
-Route::get('/getPort', [PortController::class,'getPort'] );
-Route::delete('/destroyPort/{id}', [PortController::class,'destroy'] );
+Route::get('/AdminP', [PortController::class,'portAdmin'] )->middleware('auth');
+Route::get('/getPort', [PortController::class,'getPort'] )->middleware('auth');
+Route::delete('/destroyPort/{id}', [PortController::class,'destroy'] )->middleware('auth');
 
 
 
 //ship//
 
-Route::get('/getShip', [ShipController::class,'getShip'] );
-Route::get('/getShipADD', [ShipController::class,'getShipADD'] );
-Route::get('/getcompany', [ShipController::class,'getcompany'] );
-Route::delete('/destroyShip/{id}', [ShipController::class,'destroy'] );
+Route::get('/getShip', [ShipController::class,'getShip'] )->middleware('auth');
+Route::get('/getShipADD', [ShipController::class,'getShipADD'] )->middleware('auth');
+Route::get('/getcompany', [ShipController::class,'getcompany'] )->middleware('auth');
+Route::delete('/destroyShip/{id}', [ShipController::class,'destroy'] )->middleware('auth');
 
 
 
 
 //room//
 Route::get('/getRoom_type', [RoomController::class,'getRoom_type'] );
+
+Route::get('/getRoom_Id/{roomTypeId}', [RoomController::class,'show'] );
